cleanup (using IntelliJ)

+++ src/main/java/net/sourceforge/htmlunit/xpath/functions/FuncCount.java
@@ -55,6 +55,6 @@ public class FuncCount extends FunctionOneArg {
     int i = nl.getLength();
     nl.detach();
 
-    return new XNumber((double) i);
+    return new XNumber(i);
   }
 }
